Stop using AsyncHTTPClient for testing (#20)
* Add HummingbirdCoreXCT

Which includes a basic HTTP client
Removes dependency on AsyncHTTPClient

* Separate Client.init and connect

* Create TimeoutPromise.swift

* Add connection close test and HTTP pipelining test

* swift format

* Add HTTP task to client

So I can receive response in return value of `execute`. This is more of a standard API for HTTP client.

The client assumes there will be a response for every request and does not pipeline the requests. Will need to add a timeout facility now

* Remove EventLoopStream, update tests

* tidy up

* HBHTTPClientConnection -> HBXCTClient