Remove ServerChildChannel generic associatedtype from ApplicationProtocol
Instead use any ServerChildChannel . We don’t get any real benefit from the parameter being a generic over an existential, it gets converted to an existential Service anyway. But I’ve had to jump through a number of hoops to ensure we can keep generic parameters where they do make a performance improvement. The reason I’m doing this is I was trying to flip between a TLS and non-TLS child channel when build an application and it was really quite awkward as the child channel changed the type of the Application. Removing the ServerChildChannel generic parameter simplifies this considerably.